setup coredns to route traffic internally for *.local.lco.earth
This is needed for things like generating pre-signed URLs with Minio, where
they need to use the Ingress DNS name internally. But because *.local.lco.earth
always resolves to 127.0.0.1, Pods are not able to connect to services backed
by the Ingress. By using the coredns template plugin, we can short-ciruit the
resolved DNS to the ingress-controller.